scheduler_data/requirements.txt
dbt-snowflake==1.6.0
dbt-core==1.6.0